Add ability to do in-line rule-based ignores (#2)
This adds the ability to add rule-specific in-line ignores, if you don't want to ignore the entire file.

I'm keeping it super basic intentionally. `woke` just searches for a string pattern, and if it finds it on the line, it ignores that line. Since this is a linter that can be used against any kinds of files, regardless of the language the files are written it, it has no concept of language-specific comment syntax (nor will it). Using a simple pattern matcher, `woke` can guarantee that it's in-line ignores can work on any kind of file that supports in-line comments.

Example:

```go
func main() {
  fmt.Println("here is the whitelist") // wokeignore:rule=whitelist
}
```